// src/app/generated_services/model/updateLessonDTO.ts
/**
 * CarlsonGracie Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: marcosturossi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateLessonDTO { 
    title: string;
    description?: string;
    scheduledDate?: string;
    duration?: string;
    isActive?: boolean;
}